+++ src/components/NasaCard.js
@@ -1,4 +1,4 @@
-import React from "react";
+import React, {useState, useEffect} from "react";
 
 
 const NasaCard = props => {
